Add support for an HTTP::Cookies cookie jar
This is the second half of the patch from Kari Pahula from Github:
https://github.com/kaol/HTTP-Async/commit/8e3c4d5a73f28250b515f350351dfc4876402575

Supports adding a cookie_jar per request item or on the object as a
whole. If set on the object as a whole the cookie jar is used for all
requests which are sent and responses which are received.