Hack to allow t/TestServer.pm to keep doing its proxy thing
Newer versions of HTTP::Server::Simple more correctly implement RFC1616,
which unfortunately does not leave it as a usable proxy because the
<scheme>://<host>:<port> part of the request is entirely discarded when
the request is parsed for HTTP/1.1 requests.

I've hacked in an overidden parse_request() method in TestServer.pm
which "is a" HTTP::Server::Simple::CGI, which restores the old behaviour
of not complying with RFC1616.